+++ crates/compiler/src/generate/render_resource_pots.rs
@@ -1,14 +1,15 @@
-use std::sync::Arc;
+use std::{path::Path, sync::Arc};
 
 use farmfe_core::{
   context::CompilationContext,
   error::{CompilationError, Result},
+  hashbrown::HashMap,
   parking_lot::Mutex,
   plugin::PluginHookContext,
   rayon::prelude::{IntoParallelIterator, ParallelIterator},
-  resource::{resource_pot::ResourcePot, Resource},
+  resource::{resource_pot::ResourcePot, Resource, ResourceType},
 };
-use farmfe_toolkit::fs::transform_output_filename;
+use farmfe_toolkit::fs::{transform_output_entry_filename, transform_output_filename};
 
 pub fn render_resource_pots_and_generate_resources(
   resource_pots: Vec<&mut ResourcePot>,
@@ -19,6 +20,7 @@ pub fn render_resource_pots_and_generate_resources(
   farmfe_core::puffin::profile_function!();
 
   let resources = Mutex::new(vec![]);
+  let entries = context.module_graph.read().entries.clone();
 
   // Note: Plugins should not using context.resource_pot_map, as it may cause deadlock
   resource_pots.into_par_iter().try_for_each(|resource_pot| {
@@ -33,21 +35,63 @@ pub fn render_resource_pots_and_generate_resources(
     let mut res =
       render_resource_pot_generate_resources(resource_pot, context, hook_context, false)?;
 
-    for r in &mut res {
-      let mut filename_config = context.config.output.filename.clone();
+    let mut res_map = HashMap::new();
 
-      if !r.preserve_name {
-        filename_config = transform_output_filename(
-          filename_config,
-          &r.name,
-          &r.bytes,
-          &r.resource_type.to_ext(),
-        );
-      } else {
-        filename_config = r.name.clone();
+    for r in &res {
+      // deal source map resource
+      if let ResourceType::SourceMap(original_resource_name) = &r.resource_type {
+        if let Some(orig) = res.iter().find(|item| &item.name == original_resource_name) {
+          res_map.insert(orig.name.to_string(), r.name.to_string());
+        }
       }
+    }
 
-      r.name = filename_config;
+    // deal with non-sourcemap resources
+    for r in &mut res {
+      if !matches!(r.resource_type, ResourceType::SourceMap(_)) {
+        let name_before_update = r.name.clone();
+
+        if let Some(name) = resource_pot.entry_module.as_ref() {
+          let entry_name = entries.get(name).unwrap();
+          r.name = transform_output_entry_filename(
+            context.config.output.entry_filename.clone(),
+            resource_pot.id.to_string().as_str(),
+            entry_name,
+            &r.bytes,
+            &r.resource_type.to_ext(),
+          );
+        } else {
+          r.name = transform_output_filename(
+            context.config.output.filename.clone(),
+            &r.name,
+            &r.bytes,
+            &r.resource_type.to_ext(),
+          );
+        }
+
+        if res_map.contains_key(&name_before_update) {
+          let source_mapping_url = format!(
+            "\n//# sourceMappingURL={}.{}",
+            r.name,
+            ResourceType::SourceMap("".to_string()).to_ext()
+          );
+          r.bytes.append(&mut source_mapping_url.as_bytes().to_vec());
+
+          let v = res_map.remove(&name_before_update).unwrap();
+          // reverse the map to speed up the lookup
+          res_map.insert(v, r.name.to_string());
+        }
+      }
+    }
+
+    // replace sourcemap resource
+    for r in &mut res {
+      // deal source map resource
+      if let ResourceType::SourceMap(_) = &r.resource_type {
+        if let Some(orig_name) = res_map.get(&r.name) {
+          r.name = format!("{}.{}", orig_name, r.resource_type.to_ext());
+        }
+      }
     }
 
     let mut resources = resources.lock();